backend/requirements.txt
django>=5.0,<5.2
djangorestframework>=3.15
django-cors-headers>=4.4
Pillow>=10.0
